SPEED-628 Move input package in utils
