+++ ssrf/ssrf.go
@@ -4,7 +4,6 @@ package ssrf
 import (
 	//"fmt"
 	"bytes"
-	"github.com/Contrast-Security-OSS/go-test-bench/utils/input"
 	"html/template"
 	"io/ioutil"
 	"log"
@@ -47,14 +46,14 @@ func httpHandler(w http.ResponseWriter, r *http.Request, method string) {
 	var err error
 	switch method {
 	case "query":
-		inputs := input.GetQueryInput(r, input.INPUT)
+		inputs := utils.GetQueryInput(r, utils.INPUT)
 		if inputs != "" {
 			res, err = http.Get("http://example.com?input=" + inputs)
 		} else {
 			res, err = http.Get("http://example.com")
 		}
 	case "path":
-		url := input.GetQueryInput(r, input.INPUT)
+		url := utils.GetQueryInput(r, utils.INPUT)
 		if url == "" {
 			url = "example.com"
 		}
